+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: ad3bcd36e94a9d046bc3aaa0e40af82f
-timeCreated: 1524325627
+guid: 168d40af6fcf83845b4aad38b57b7534
+folderAsset: yes
+timeCreated: 1524329394
 licenseType: Free
-NativeFormatImporter:
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
