+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/NodeEditorBackground.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0e1db3de822a32744aeadcba119135e4
+timeCreated: 1524329974
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
