+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/NodeEditorLinks.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: f8d77b289869b714e9c851a61871dd2b
+timeCreated: 1524330360
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
