+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/NodeEditorNodes.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 3513b2e1775c95748b44105b7a5ba9d2
+timeCreated: 1524347669
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
