+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/NodeEditorPanel.cs.meta
@@ -1,8 +1,9 @@
 fileFormatVersion: 2
-guid: 2a10edc594f7f8d48a48310ba1f1b248
-timeCreated: 1509291289
+guid: a52b02ce9e39feb40b726aaeb0741fc1
+timeCreated: 1524329394
 licenseType: Free
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
