+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/interfaces.meta
@@ -1,10 +1,10 @@
 fileFormatVersion: 2
-guid: ad3bcd36e94a9d046bc3aaa0e40af82f
-timeCreated: 1524325627
+guid: 2522d66f0a7f43540885ce1da1a0abf5
+folderAsset: yes
+timeCreated: 1524342130
 licenseType: Free
-NativeFormatImporter:
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
