+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/interfaces/ILinkEditor.cs
@@ -1,0 +1,8 @@
+using Constellation;
+
+namespace ConstellationEditor {
+    public interface ILinkEditor {
+        void AddLinkFromInput(InputData selectedInput);
+        void AddLinkFromOutput(OutputData selectedOutput);
+    }
+}
