+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/interfaces/ILinkEditor.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 22791fae5fa63b94ca288b76ee3a0085
+timeCreated: 1524342130
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
