+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/interfaces/IVisibleObject.cs
@@ -1,0 +1,6 @@
+using UnityEngine;
+namespace ConstellationEditor {
+    public interface IVisibleObject {
+        bool InView (Rect rect);
+    }
+}
