+++ Constellation/Assets/Constellation/Editor/NodeEditor/NodeEditorPannel/interfaces/IVisibleObject.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: d4ed91b1053e19c49bfda641cdb1a99b
+timeCreated: 1524363828
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
